迁移 MiotEventTrigger schema 至 miot 包消除循环导入，删除 translations 手写翻译表改用 spec value_translate
